requirements.txt
pyyaml==6.0.2
chromadb==1.0.12
fastapi==0.115.9
uvicorn==0.34.3